+++ Dialogix_J2EE/DialogixEntities/src/java/org/dialogix/entities/Answer.java
@@ -10,7 +10,7 @@
 package org.dialogix.entities;
 
 import java.io.Serializable;
-import java.math.BigInteger;
+
 import java.util.Collection;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
@@ -30,7 +30,7 @@ public class Answer implements Serializable {
     @Id
     @GeneratedValue(strategy=GenerationType.TABLE, generator="Answer_Gen")
     @Column(name = "answer_id", nullable = false)
-    private BigInteger answerID;
+    private Long answerID;
     @Column(name = "hasLAcode")
     private Boolean hasLAcode;
     @Column(name = "LAcode")
@@ -49,15 +49,15 @@ public class Answer implements Serializable {
     public Answer() {
     }
 
-    public Answer(BigInteger answerID) {
+    public Answer(Long answerID) {
         this.answerID = answerID;
     }
 
-    public BigInteger getAnswerID() {
+    public Long getAnswerID() {
         return answerID;
     }
 
-    public void setAnswerID(BigInteger answerID) {
+    public void setAnswerID(Long answerID) {
         this.answerID = answerID;
     }
