+++ Dialogix_J2EE/DialogixEntities/src/java/org/dialogix/entities/InstrumentContent.java
@@ -10,7 +10,7 @@
 package org.dialogix.entities;
 
 import java.io.Serializable;
-import java.math.BigInteger;
+
 import java.util.Collection;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
@@ -33,7 +33,7 @@ public class InstrumentContent implements Serializable {
     @Id
     @GeneratedValue(strategy=GenerationType.TABLE, generator="InstrumentContent_Gen")
     @Column(name = "instrument_content_id", nullable = false)
-    private BigInteger instrumentContentID;
+    private Long instrumentContentID;
     @Column(name = "ItemSequence", nullable = false)
     private int itemSequence;
     @Lob
@@ -93,11 +93,11 @@ public class InstrumentContent implements Serializable {
     public InstrumentContent() {
     }
 
-    public InstrumentContent(BigInteger instrumentContentID) {
+    public InstrumentContent(Long instrumentContentID) {
         this.instrumentContentID = instrumentContentID;
     }
 
-    public InstrumentContent(BigInteger instrumentContentID, int itemSequence, int isRequired, int isReadOnly, int groupNum, String relevance, int isMessage) {
+    public InstrumentContent(Long instrumentContentID, int itemSequence, int isRequired, int isReadOnly, int groupNum, String relevance, int isMessage) {
         this.instrumentContentID = instrumentContentID;
         this.itemSequence = itemSequence;
         this.isRequired = isRequired;
@@ -107,11 +107,11 @@ public class InstrumentContent implements Serializable {
         this.isMessage = isMessage;
     }
 
-    public BigInteger getInstrumentContentID() {
+    public Long getInstrumentContentID() {
         return instrumentContentID;
     }
 
-    public void setInstrumentContentID(BigInteger instrumentContentID) {
+    public void setInstrumentContentID(Long instrumentContentID) {
         this.instrumentContentID = instrumentContentID;
     }
